@@ -1,16 +1,15 @@
 name: date_picker_timeline
 description: Flutter Date Picker Library that provides a calendar as a horizontal timeline.
-version: 1.1.3
-author: Vivek Kaushik<me@vivekkaushik.com>
+version: 1.2.3
 homepage: https://github.com/iamvivekkaushik/DatePickerTimelineFlutter
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  intl: ^0.16.0
+  intl: ^0.17.0
 
 dev_dependencies:
   flutter_test:
@@ -38,13 +37,13 @@ flutter:
   # "family" key with the font family name, and a "fonts" key with a
   # list giving the asset and other descriptors for the font. For
   # example:
-   fonts:
-     - family: Roboto
-       fonts:
-         - asset: fonts/RobotoRegular.ttf
-           weight: 500
-         - asset: fonts/RobotoBold.ttf
-           weight: 700
+  #   fonts:
+  #   - family: Roboto
+  #     fonts:
+  #       - asset: fonts/RobotoRegular.ttf
+  #         weight: 500
+  #       - asset: fonts/RobotoBold.ttf
+  #         weight: 700
   #   - family: Schyler
   #     fonts:
   #       - asset: fonts/Schyler-Regular.ttf
